requirements.txt
ImageHash==4.3.1
pillow==10.2.0
python-dotenv==1.0.1
python-telegram-bot[job-queue]==20.8
PyYAML==6.0.1
telegram==0.0.1
ujson==5.9.0
